+++ backend/src/main/java/ru/shutovna/moyserf/repository/WalletRepository.java
@@ -3,5 +3,5 @@ package ru.shutovna.moyserf.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.shutovna.moyserf.model.Wallet;
 
-public interface WalletRepository extends JpaRepository<Wallet, Long> {
+public interface WalletRepository extends JpaRepository<Wallet, Integer> {
 }
