+++ backend/src/main/java/ru/shutovna/moyserf/service/IAvatarService.java
@@ -14,5 +14,5 @@ public interface IAvatarService {
 
     void rotateAvatar(double angle);
 
-    UrlResource getAvatar(long userId) throws MalformedURLException;
+    UrlResource getAvatar(int userId) throws MalformedURLException;
 }
